better isbns

+++ scripts/collect-isbns.py
@@ -48,7 +48,7 @@ isbns = loc_isbns.join(ol_isbns, how='outer')
 isbns = isbns.join(gr_isbns, how='outer')
 isbns = isbns.join(bx_isbns, how='outer')
 isbns.sort_index(inplace=True)
-isbns = isbns.fillna(0).astype('i32')
+isbns = isbns.fillna(0).astype('i4')
 _log.info('found %d unique ISBNs', len(isbns))
 isbns['isbn_id'] = np.arange(len(isbns), dtype=np.int32) + 1
 isbns.index.name = 'isbn'
